backend/requirements.txt
flask>=2.0.0
flask-cors==4.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
dnspython==2.6.1
flask-jwt-extended>=4.4.0
bcrypt==4.1.2
PyJWT==2.8.0
validators==0.22.0
python-jose==3.3.0
sendgrid==6.11.0
python-decouple==3.8